+++ src/main/java/com/dateplanner/admin/user/dto/UserJoinRequestDto.java
@@ -1,6 +1,6 @@
-package com.dateplanner.user.dto;
+package com.dateplanner.admin.user.dto;
 
-import com.dateplanner.user.entity.User;
+import com.dateplanner.admin.user.entity.User;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AccessLevel;
 import lombok.Builder;
