+++ src/main/java/com/dateplanner/admin/user/dto/UserRequestDto.java
@@ -1,6 +1,6 @@
-package com.dateplanner.user.dto;
+package com.dateplanner.admin.user.dto;
 
-import com.dateplanner.user.entity.User;
+import com.dateplanner.admin.user.entity.User;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -14,11 +14,14 @@ import javax.validation.constraints.NotNull;
 @Getter
 @Setter
 @NoArgsConstructor
-public class UserModifyRequestDto {
+public class UserRequestDto {
 
     /**
-     * 회원 수정용 Dto
+     * 회원 생성용 Dto
      */
+
+    @NotNull
+    private String uid;
 
     @NotNull
     private String password;
@@ -32,9 +35,20 @@ public class UserModifyRequestDto {
     private String introduce;
 
     @Builder
-    public UserModifyRequestDto(String password, String email, String introduce) {
+    public UserRequestDto(String uid, String password, String email, String introduce) {
+        this.uid = uid;
         this.password = password;
         this.email = email;
         this.introduce = introduce;
     }
+
+    public User toEntity() {
+        return User.builder()
+                .uid(uid)
+                .password(password)
+                .email(email)
+                .introduce(introduce)
+                .build();
+    }
+
 }
