+++ src/main/java/com/dateplanner/admin/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.dateplanner.user.repository;
+package com.dateplanner.admin.user.repository;
 
-import com.dateplanner.user.entity.User;
+import com.dateplanner.admin.user.entity.User;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
