+++ src/main/java/com/dateplanner/bookmark/service/BookmarkService.java
@@ -7,8 +7,8 @@ import com.dateplanner.bookmark.entity.Bookmark;
 import com.dateplanner.bookmark.repository.BookmarkRepository;
 import com.dateplanner.place.entity.Place;
 import com.dateplanner.place.repository.PlaceRepository;
-import com.dateplanner.user.entity.User;
-import com.dateplanner.user.repository.UserRepository;
+import com.dateplanner.admin.user.entity.User;
+import com.dateplanner.admin.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
