+++ src/main/java/com/dateplanner/plan/entity/Plan.java
@@ -1,7 +1,7 @@
 package com.dateplanner.plan.entity;
 
 import com.dateplanner.constant.entity.BaseTimeEntity;
-import com.dateplanner.user.entity.User;
+import com.dateplanner.admin.user.entity.User;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
