+++ src/main/java/com/dateplanner/plan/service/PlanService.java
@@ -6,8 +6,8 @@ import com.dateplanner.plan.dto.PlanRequestDto;
 import com.dateplanner.plan.entity.Plan;
 import com.dateplanner.plan.repository.DetailPlanRepository;
 import com.dateplanner.plan.repository.PlanRepository;
-import com.dateplanner.user.entity.User;
-import com.dateplanner.user.repository.UserRepository;
+import com.dateplanner.admin.user.entity.User;
+import com.dateplanner.admin.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
