+++ src/main/java/com/dateplanner/review/service/ReviewService.java
@@ -8,8 +8,8 @@ import com.dateplanner.place.repository.PlaceRepository;
 import com.dateplanner.review.dto.ReviewRequestDto;
 import com.dateplanner.review.entity.Review;
 import com.dateplanner.review.repository.ReviewRepository;
-import com.dateplanner.user.entity.User;
-import com.dateplanner.user.repository.UserRepository;
+import com.dateplanner.admin.user.entity.User;
+import com.dateplanner.admin.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
