+++ src/test/java/com/dateplanner/bookmark/service/BookmarkServiceTest.java
@@ -5,8 +5,8 @@ import com.dateplanner.bookmark.repository.BookmarkRepository;
 import com.dateplanner.fixture.Fixture;
 import com.dateplanner.place.entity.Place;
 import com.dateplanner.place.repository.PlaceRepository;
-import com.dateplanner.user.entity.User;
-import com.dateplanner.user.repository.UserRepository;
+import com.dateplanner.admin.user.entity.User;
+import com.dateplanner.admin.user.repository.UserRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
