+++ src/test/java/com/dateplanner/plan/service/PlanServiceTest.java
@@ -5,8 +5,8 @@ import com.dateplanner.plan.dto.PlanRequestDto;
 import com.dateplanner.plan.entity.Plan;
 import com.dateplanner.plan.repository.DetailPlanRepository;
 import com.dateplanner.plan.repository.PlanRepository;
-import com.dateplanner.user.entity.User;
-import com.dateplanner.user.repository.UserRepository;
+import com.dateplanner.admin.user.entity.User;
+import com.dateplanner.admin.user.repository.UserRepository;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
